+++ compiler/src/Type/Comparable.hs
@@ -61,7 +61,7 @@ data Info =
 
 
 compute :: Map.Map ModuleName.Raw I.Interface -> Can.Module -> Info
-compute ifaces (Can.Module home _ _ _ unions _ _ _) =
+compute ifaces (Can.Module home _ _ _ unions _ _ _ _) =
   let
     imported =
       Set.unions (map I._comparables (Map.elems ifaces))
@@ -152,6 +152,9 @@ isComparableType home imported unions seen tipe =
       False
 
     Can.TRecord _ _ ->
+      False
+
+    Can.TTagRow _ _ ->
       False