+++ terminal/src/Repl.hs
@@ -333,10 +333,11 @@ attemptDeclOrExpr lines =
       case dResult of
         Right (decl, _) ->
           case decl of
-            PD.Value _ (A.At _ (Src.Value (A.At _ name) _ _ _)) -> ifDone lines (Decl name src)
-            PD.Union _ (A.At _ (Src.Union (A.At _ name) _ _  )) -> ifDone lines (Type name src)
-            PD.Alias _ (A.At _ (Src.Alias (A.At _ name) _ _  )) -> ifDone lines (Type name src)
-            PD.Port  _ _                                        -> pure $ Done Port
+            PD.Value _ (A.At _ (Src.Value (A.At _ name) _ _ _))    -> ifDone lines (Decl name src)
+            PD.Union _ (A.At _ (Src.Union (A.At _ name) _ _  ))    -> ifDone lines (Type name src)
+            PD.Alias _ (A.At _ (Src.Alias (A.At _ name) _ _  ))    -> ifDone lines (Type name src)
+            PD.TagDecl _ (A.At _ (Src.TagDecl (A.At _ name) _))    -> ifDone lines (Type name src)
+            PD.Port  _ _                                           -> pure $ Done Port
 
         Left declPosition
           | startsWithKeyword "type" lines -> ifFail lines (Type "ERR" src)